requirements.txt
boto3==1.24.89
botocore==1.27.89
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==1.15.0
dnspython3==1.15.0
idna==3.4
jmespath==1.0.1
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-lambda-local==0.1.13
pyUML==0.1.0a0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
pytest==7.2.0
pytest-mock==3.10.0
coverage==6.5.0
fpdf~=1.7.2
igraph~=0.10.4
Jinja2~=3.1.2
dataclasses~=0.6
minio~=7.1.13
squiggle~=0.3.1
numpy~=1.24.2
pandas~=1.5.3
tqdm~=4.64.1